Python/web/auth_app/flaskr/config.py
import os
from dotenv import load_dotenv


load_dotenv(override=True)


class Development:
    DEBUG = True
    SECRET_KEY=os.getenv('SECRET_KEY')
    SQLALCHEMY_DATABASE_URI=os.getenv('SQLALCHEMY_DATABASE_URI')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SQLALCHEMY_ECHO = False
